tm dbhelper part

+++ app/src/main/java/com/comp3617/finalproject/meggsage/TextReminder.java
@@ -30,7 +30,7 @@ public class TextReminder {
         return createDate;
     }
 
-    public void setCreateDate(int createDate) {
+    public void setCreateDate(long createDate) {
         this.createDate = createDate;
     }
